chore(git): migrate inline tests to separated files and fix rebase (#530)
Migrate all 9 #530 crates to comply with #569 test layout policy:
- driver-git: provider_tests.rs, store_tests.rs, types_tests.rs
- module-git: tests.rs, subprocess_tests.rs
- picker-git-{branches,log,stash,status}: tests.rs

Fix rebase compatibility issues:
- Add missing PickerContext `options` field in 4 git picker crates
- Fix flatten() call signature in explorer file_tree_tests.rs
- Add layer opacity test and update window binding count (27 -> 29)

Refs #530
